requirements.txt
python-telegram-bot>=21.6
requests>=2.32
python-dotenv>=1.0
matplotlib>=3.8